Restructure tests
It became more and more difficult to find a good place to put new tests.
Hence new tests division is introduced. Basic idea is to have a separate
file for testing every feature.

Commit contains tests move and some minor changes (not influencing tests
behavior), usually in form of renaming and moving test data files to
better reflect new structure.

General move rules (singular tests may break it):

Tests from test_core are moved to test_helpers
Tests from test_dlis are mainly moved to test_logical_file and
test_basic_object
Tests from test_fdata are moved to test_curves_reprcodes and test_curves
Tests from test_frames are moved to test_curves
Tests from test_layout are moved to test_physical_layout
Tests from test_load are mainly moved to test_loading and
test_partitioning
Tests from test_parse are moved to test_logical_record and partly to
test_encodings
Tests from test_semantics are mainly moved to test_object_structures
Tests from test_tweaking are moved to test_monkey_patching
Tests from test_utils are mainly moved to test_basic_object and
test_dimensional_attributes
