+++ python/data/chap2/README.rst
@@ -8,34 +8,23 @@ Visible Records, Storage Unit Labels, padbytes and garbage data.
 =================================== ===========================================
 File                                Description
 =================================== ===========================================
+1lr-in-2vrs.dlis                    A logical record spanning 2 visible records
+
 2lr-in-vr.dlis                      2 logical records in 1 visible record
+
+3lrs-in-vr.dlis                     A single visible record with one logical
+                                    record segmented into 3 logical record
+                                    segments
+
+3lr-in-vr-one-encrypted.dlis        Single VR with with 3 logical records, one
+                                    encrypted
 
 7K-file.dlis                        Testing that files between 4 and 8 kB are
                                     read correctly.
 
-example-record.dlis                 A single visible record with one logical
-                                    record segmented into 3 logical record
-                                    segments
-
-fdata-encrypted.dlis                One encrypted fdata with characters not in
-                                    utf-8 set
-
-fdata-many-in-same-vr.dlis          A single visible record, with several
-                                    Indirectly formatted logical records of type
-                                    fdata
-
-fdata-vr-aligned.dlis               One fdata aligned with visible record
-
-fdata-vr-disaligned-in-obname.dlis  One fdata disaligned with visible record,
-                                    but cut happens inside obname
-
-fdata-vr-disaligned.dlis            One fdata disaligned with visible record
-
 incomplete-sul.dlis                 Incomplete storage unit label
 
 incomplete-vr.dlis                  Incomplete visible record
-
-lr-in-2vrs.dlis                     A logical record spanning 2 visible records
 
 nondlis.txt                         Not a dlis file
 
@@ -57,6 +46,7 @@ pre-sul-garbage.dlis                Garbage before the storage unit label
 
 pre-sul-pre-vrl-garbage.dlis        Garbage before storage unit label and
                                     visible record
+
 small.dlis                          Test file smaller than 4kB
 
 too-small-record.dlis               Visible record is smaller than the minimum
