+++ python/data/chap3/implicit/fdata-encrypted.dlis
