+++ python/tests/test_curves.py
@@ -1,45 +1,20 @@
-import numpy as np
+"""
+Testing frame.curves() and channel.curves() functions and all the
+helper functionality
+"""
 import pytest
+import numpy as np
+from datetime import datetime
 
 import dlisio
+from dlisio import core
 from dlisio.plumbing import mkunique
 
-from dlisio import core
-
-def test_curves_values(f):
-    frame = f.object('FRAME', 'FRAME1', 10, 0)
-    curves = frame.curves()
-
-    base = np.array([
-        [[1, 2], [3, 4], [5, 6]],
-        [[7, 8], [9, 10], [11, 12]],
-        [[13, 14], [15, 16], [17, 18]],
-        [[19, 20], [21, 22], [23, 24]]
-    ])
-
-    val = base + 0
-    np.testing.assert_array_equal(curves['CHANN1'][0], val)
-    np.testing.assert_array_equal(curves[0]['CHANN1'], val)
-
-    val = base + 256
-    np.testing.assert_array_equal(curves['CHANN2'][0], val)
-    np.testing.assert_array_equal(curves[0]['CHANN2'], val)
-
-    val = base + 512
-    np.testing.assert_array_equal(curves['CHANN1'][1], val)
-    np.testing.assert_array_equal(curves[1]['CHANN1'], val)
-
-    val = base + 768
-    np.testing.assert_array_equal(curves['CHANN2'][1], val)
-    np.testing.assert_array_equal(curves[1]['CHANN2'], val)
-
-    val = base + 1024
-    np.testing.assert_array_equal(curves['CHANN1'][2], val)
-    np.testing.assert_array_equal(curves[2]['CHANN1'], val)
-
-    val = base + 1280
-    np.testing.assert_array_equal(curves['CHANN2'][2], val)
-    np.testing.assert_array_equal(curves[2]['CHANN2'], val)
+def load_curves(fpath):
+    with dlisio.load(fpath) as (f, *_):
+        frame = f.object('FRAME', 'FRAME-REPRCODE', 10, 0)
+        curves = frame.curves()
+        return curves
 
 def makeframe():
     frame = dlisio.plumbing.Frame()
@@ -97,6 +72,141 @@ def makeframe():
 
     frame.link()
     return frame
+
+def test_curves_values(f):
+    frame = f.object('FRAME', 'FRAME1', 10, 0)
+    curves = frame.curves()
+
+    base = np.array([
+        [[1, 2],   [3, 4],   [5, 6]],
+        [[7, 8],   [9, 10],  [11, 12]],
+        [[13, 14], [15, 16], [17, 18]],
+        [[19, 20], [21, 22], [23, 24]]
+    ])
+
+    val = base + 0
+    np.testing.assert_array_equal(curves['CHANN1'][0], val)
+    np.testing.assert_array_equal(curves[0]['CHANN1'], val)
+
+    val = base + 256
+    np.testing.assert_array_equal(curves['CHANN2'][0], val)
+    np.testing.assert_array_equal(curves[0]['CHANN2'], val)
+
+    val = base + 512
+    np.testing.assert_array_equal(curves['CHANN1'][1], val)
+    np.testing.assert_array_equal(curves[1]['CHANN1'], val)
+
+    val = base + 768
+    np.testing.assert_array_equal(curves['CHANN2'][1], val)
+    np.testing.assert_array_equal(curves[1]['CHANN2'], val)
+
+    val = base + 1024
+    np.testing.assert_array_equal(curves['CHANN1'][2], val)
+    np.testing.assert_array_equal(curves[2]['CHANN1'], val)
+
+    val = base + 1280
+    np.testing.assert_array_equal(curves['CHANN2'][2], val)
+    np.testing.assert_array_equal(curves[2]['CHANN2'], val)
+
+def test_two_various_fdata_in_one_iflr():
+    fpath = 'data/chap4-7/iflr/two-various-fdata-in-one-iflr.dlis'
+
+    curves = load_curves(fpath)
+    assert curves[0][1] == datetime(1971, 3, 21, 18, 4, 14, 386000)
+    assert curves[0][2] == "VALUE"
+    assert curves[0][3] == 89
+    assert curves[1][1] == datetime(1970, 3, 21, 18, 4, 14, 0)
+    assert curves[1][2] == "SECOND-VALUE"
+    assert curves[1][3] == -89
+
+def test_out_of_order_framenos_one_frame():
+    fpath = 'data/chap4-7/iflr/out-of-order-framenos-one-frame.dlis'
+    curves = load_curves(fpath)
+    np.testing.assert_array_equal(curves['FRAMENO'], [2, 1])
+    assert curves[0][1] == True
+    assert curves[1][1] == False
+
+def test_out_of_order_framenos_two_frames():
+    fpath = 'data/chap4-7/iflr/out-of-order-framenos-two-frames.dlis'
+    curves = load_curves(fpath)
+    np.testing.assert_array_equal(curves['FRAMENO'], [2, 1])
+    assert curves[0][1] == True
+    assert curves[1][1] == False
+
+def test_out_of_order_frames_two_framenos_multidata():
+    fpath = 'data/chap4-7/iflr/out-of-order-framenos-two-frames-multifdata.dlis'
+    curves = load_curves(fpath)
+    np.testing.assert_array_equal(curves['FRAMENO'], [3, 4, 1, 2])
+    assert curves[0][1] == True
+    assert curves[1][1] == False
+    assert curves[2][1] == False
+    assert curves[3][1] == True
+
+def test_missing_numbers_frames():
+    fpath = 'data/chap4-7/iflr/missing-framenos.dlis'
+    curves = load_curves(fpath)
+    np.testing.assert_array_equal(curves['FRAMENO'], [2, 4])
+    assert curves[0][1] == True
+    assert curves[1][1] == False
+
+def test_duplicate_framenos():
+    fpath = 'data/chap4-7/iflr/duplicate-framenos.dlis'
+    curves = load_curves(fpath)
+    assert curves[0][0] == curves[0][1]
+    assert curves[0][0] == True
+    assert curves[1][1] == False
+
+def test_duplicate_framenos_same_frames():
+    fpath = 'data/chap4-7/iflr/duplicate-framenos-same-frames.dlis'
+    curves = load_curves(fpath)
+    np.testing.assert_array_equal(curves[0], curves[1])
+
+def test_fdata_dimension():
+    fpath = 'data/chap4-7/iflr/multidimensions-ints-various.dlis'
+
+    with dlisio.load(fpath) as (f, *_):
+        frame = f.object('FRAME', 'FRAME-DIMENSION', 11, 0)
+        curves = frame.curves()
+
+        np.testing.assert_array_equal(curves[0][1], [[1, 2, 3], [4, 5, 6]])
+        np.testing.assert_array_equal(curves[0][2], [[1, 2], [3, 4], [5, 6]])
+
+        arr2 = [
+            [[1, 2],   [3, 4],   [5, 6]],
+            [[7, 8],   [9, 10],  [11, 12]],
+            [[13, 14], [15, 16], [17, 18]],
+            [[19, 20], [21, 22], [23, 24]]
+        ]
+        np.testing.assert_array_equal(curves[0][3], arr2)
+        np.testing.assert_array_equal(curves[0][4], [[1, 2]])
+        np.testing.assert_array_equal(curves[0][5], [[1], [2]])
+        np.testing.assert_array_equal(curves[0][6], [[1]])
+        np.testing.assert_array_equal(curves[0][7], [1, 2, 3, 4])
+
+def test_fdata_tuple_dimension():
+    fpath = 'data/chap4-7/iflr/multidimensions-validated.dlis'
+
+    with dlisio.load(fpath) as (f, *_):
+        frame = f.object('FRAME', 'FRAME-VALIDATE', 10, 0)
+        curves = frame.curves()
+
+        assert curves[0][1].size == 3
+
+        assert curves[0][1][0] == (56, 0.0625, 0.0625)
+        assert curves[0][1][1] == (43, 0.0625, 0.0625)
+        assert curves[0][1][2] == (71, 0.5, 0.5)
+
+def test_fdata_dimensions_in_multifdata():
+    fpath = 'data/chap4-7/iflr/multidimensions-multifdata.dlis'
+
+    with dlisio.load(fpath) as (f, *_):
+        frame = f.object('FRAME', 'FRAME-DIMENSION', 11, 0)
+        curves = frame.curves()
+
+        assert curves.shape == (2,)
+
+        np.testing.assert_array_equal(curves[0][1], [[1, 2, 3], [4, 5, 6]])
+        np.testing.assert_array_equal(curves[1][1], [[7, 8, 9], [10, 11, 12]])
 
 def test_duplicated_mnemonics_gets_unique_labels():
     frame = makeframe()
@@ -181,18 +291,6 @@ def test_instance_dtype_fmt():
     expected_names = ('FRAMENO', 'x-TIME 0~0', 'TDEP', 'x-TIME 1~0')
     assert expected_names == frame.dtype().names
 
-@pytest.mark.parametrize('fmt', [
-    ("x-{:d}.{:s}.{:d}"),
-    ("x-{:s}.{:d}.{:d}.{:d}"),
-])
-def test_instance_dtype_wrong_fmt(fmt, assert_log):
-    frame = makeframe()
-
-    frame.dtype_fmt = fmt
-    with pytest.raises(Exception):
-        _ = frame.dtype().names
-    assert_log("rich label")
-
 def test_class_dtype_fmt():
     original = dlisio.plumbing.Frame.dtype_format
 
@@ -209,58 +307,86 @@ def test_class_dtype_fmt():
         # default, to not interfere with other tests
         dlisio.plumbing.Frame.dtype_format = original
 
+@pytest.mark.parametrize('fmt', [
+    ("x-{:d}.{:s}.{:d}"),
+    ("x-{:s}.{:d}.{:d}.{:d}"),
+])
+def test_instance_dtype_wrong_fmt(fmt, assert_log):
+    frame = makeframe()
+
+    frame.dtype_fmt = fmt
+    with pytest.raises(Exception):
+        _ = frame.dtype().names
+    assert_log("rich label")
+
+
 def test_channel_curves():
+    fpath = 'data/chap4-7/iflr/all-reprcodes.dlis'
+    with dlisio.load(fpath) as (f, *_):
+        channel = f.object('CHANNEL', 'CH26', 10, 0)
+        curves = channel.curves()
+        assert curves[0] == True
+
+        channel = f.object('CHANNEL', 'CH22', 10, 0)
+        curves22 = channel.curves()
+        assert curves22[0] == 16777217
+
+        frame_curves = load_curves(fpath)
+        assert frame_curves['CH22'] == curves22
+
+
+def test_channel_curves_v2():
     ch1 = dlisio.plumbing.Channel()
-    ch1.name       ='ch1'
-    ch1.origin     = 0
+    ch1.name = 'ch1'
+    ch1.origin = 0
     ch1.copynumber = 0
     ch1.attic = {
-        'DIMENSION'           : [5],
-        'REPRESENTATION-CODE' : [11],
+        'DIMENSION': [5],
+        'REPRESENTATION-CODE': [11],
     }
 
     ch2 = dlisio.plumbing.Channel()
-    ch2.name       ='ch2'
-    ch2.origin     = 0
+    ch2.name = 'ch2'
+    ch2.origin = 0
     ch2.copynumber = 0
     ch2.attic = {
-        'DIMENSION'           : [2, 2],
-        'REPRESENTATION-CODE' : [3],
+        'DIMENSION': [2, 2],
+        'REPRESENTATION-CODE': [3],
     }
 
     ch3 = dlisio.plumbing.Channel()
-    ch3.name       ='ch3'
-    ch3.origin     = 0
+    ch3.name = 'ch3'
+    ch3.origin = 0
     ch3.copynumber = 0
     ch3.attic = {
-        'DIMENSION'           : [4, 2],
-        'REPRESENTATION-CODE' : [26],
+        'DIMENSION': [4, 2],
+        'REPRESENTATION-CODE': [26],
     }
 
     ch4 = dlisio.plumbing.Channel()
-    ch4.name       ='ch4'
-    ch4.origin     = 0
+    ch4.name = 'ch4'
+    ch4.origin = 0
     ch4.copynumber = 0
     ch4.attic = {
-        'DIMENSION'           : [1],
-        'REPRESENTATION-CODE' : [17],
+        'DIMENSION': [1],
+        'REPRESENTATION-CODE': [17],
     }
 
     ch5 = dlisio.plumbing.Channel()
-    ch5.name       ='ch5'
-    ch5.origin     = 0
+    ch5.name = 'ch5'
+    ch5.origin = 0
     ch5.copynumber = 0
     ch5.attic = {
-        'DIMENSION'           : [2, 3, 1],
-        'REPRESENTATION-CODE' : [12],
+        'DIMENSION': [2, 3, 1],
+        'REPRESENTATION-CODE': [12],
     }
 
     frame = dlisio.plumbing.Frame()
-    frame.name       ='fr'
-    frame.origin     = 0
+    frame.name = 'fr'
+    frame.origin = 0
     frame.copynumber = 0
     frame.attic = {
-        'CHANNELS' : [
+        'CHANNELS': [
             core.obname(ch1.origin, ch1.copynumber, ch1.name),
             core.obname(ch2.origin, ch2.copynumber, ch2.name),
             core.obname(ch3.origin, ch3.copynumber, ch3.name),
@@ -271,34 +397,35 @@ def test_channel_curves():
 
     logicalfile = dlisio.dlis(None, [], [], [])
     logicalfile.indexedobjects['FRAME'] = {
-        frame.fingerprint : frame
+        frame.fingerprint: frame
     }
 
     logicalfile.indexedobjects['CHANNEL'] = {
-        ch1.fingerprint : ch1,
-        ch2.fingerprint : ch2,
-        ch3.fingerprint : ch3,
-        ch4.fingerprint : ch4,
-        ch5.fingerprint : ch5,
+        ch1.fingerprint: ch1,
+        ch2.fingerprint: ch2,
+        ch3.fingerprint: ch3,
+        ch4.fingerprint: ch4,
+        ch5.fingerprint: ch5,
     }
     frame.logicalfile = logicalfile
-    ch1.logicalfile   = logicalfile
-    ch2.logicalfile   = logicalfile
-    ch3.logicalfile   = logicalfile
-    ch4.logicalfile   = logicalfile
-    ch5.logicalfile   = logicalfile
+    ch1.logicalfile = logicalfile
+    ch2.logicalfile = logicalfile
+    ch3.logicalfile = logicalfile
+    ch4.logicalfile = logicalfile
+    ch5.logicalfile = logicalfile
 
     pre_fmt, ch_fmt, post_fmt = frame.fmtstrchannel(ch3)
-    assert pre_fmt  == "CCCCCbbbb"
-    assert ch_fmt   == "qqqqqqqq"
+    assert pre_fmt == "CCCCCbbbb"
+    assert ch_fmt == "qqqqqqqq"
     assert post_fmt == "Ldddddd"
+
 
 def test_channel_no_dimension(assert_log):
     ch = dlisio.plumbing.Channel()
     ch.name = 'CH'
     ch.origin = 0
     ch.copynumber = 0
-    ch.attic = { 'REPRESENTATION-CODE' : [17] }
+    ch.attic = {'REPRESENTATION-CODE': [17]}
 
     with pytest.raises(ValueError) as exc:
         ch.fmtstr()
@@ -306,6 +433,7 @@ def test_channel_no_dimension(assert_log):
 
     ch.attic['DIMENSION'] = [1]
     assert ch.fmtstr() == "L"
+
 
 def test_frame_index():
     frame = makeframe()
@@ -323,3 +451,4 @@ def test_frame_nochannels_no_index(assert_info):
 
     assert frame.index is None
     assert_info('Frame has no channels')
+
