+++ python/tests/test_logical_file.py
@@ -1,3 +1,7 @@
+"""
+Testing Logical file class (also known as dlis)
+"""
+
 import dlisio
 import pytest
 
@@ -9,7 +13,7 @@ from dlisio import core
 
 @pytest.fixture(scope="module")
 def g():
-    s = dlisio.open("tests/test_dlis.py") #any existing file is required
+    s = dlisio.open("tests/test_logical_file.py") #any existing file is required
     g = dlisio.dlis(s, [], [], [])
 
     ch = Channel()
@@ -175,42 +179,6 @@ def test_match_special_characters(g):
     for ch in channels:
         assert ch in refs
 
-def test_basicobject_getitem_defaultvalue(f):
-    obj = f.object('FRAME', 'FRAME2')
-
-    # Attribute index_type is absent (from attic), but we stil expect to get a
-    # default value
-    assert obj['INDEX-TYPE'] is None
-
-def test_basicobject_getitem_unexpected_attr(f):
-    obj = f.object('FRAME', 'FRAME2')
-
-    try:
-        obj.attic['NEW-ATTR'] = [1]
-
-        # Attributes unknown to dlisio, such as 'NEW-ATTR' should be reachable
-        # through __getitem__
-        assert obj['NEW-ATTR'] == [1]
-
-        # Should also be in stash
-        assert obj.stash['NEW-ATTR'] == [1]
-    finally:
-        del obj.attic['NEW-ATTR']
-
-def test_basicobject_getitem_noattribute(f):
-    obj = f.object('FRAME', 'FRAME2')
-
-    # getitem should raise an KeyError if key not in obj.attic or
-    # obj.attributes
-
-    with pytest.raises(KeyError):
-        _ = obj['DUMMY']
-
-def test_basicobject_getitem(f):
-    obj = f.object('FRAME', 'FRAME1')
-
-    assert obj['INDEX-TYPE'] == 'BOREHOLE-DEPTH'
-
 def test_indexedobjects(f):
     assert f.fileheader.name   == "N"
     assert len(f.origins)      == 2
@@ -238,6 +206,18 @@ def test_initial_load(fpath):
     with dlisio.load(fpath) as (f, *tail):
         # Only fileheader, origin, frame, and channel should be loaded
         assert len(f.indexedobjects) == 4
+
+
+def test_load_all(fpath):
+    with dlisio.load(fpath) as (f, *_):
+        f.load()
+        assert len(f.indexedobjects) == 23
+
+def test_unknowns_loading():
+    with dlisio.load('data/206_05a-_3_DWL_DWL_WIRE_258276498.DLIS') as (f,):
+        assert len(f.indexedobjects) == 4 #FILE-HEADER, ORIGIN, FRAME, CHANNEL
+        assert len(f.unknowns)       == 5
+        assert len(f.indexedobjects) == 9
 
 def test_objecttype_load(fpath):
     with dlisio.load(fpath) as (f, *tail):
@@ -271,14 +251,3 @@ def test_object_load_link(fpath):
         _ = ch.long_name
         assert fp in f.indexedobjects['LONG-NAME']
         assert len(f.indexedobjects['LONG-NAME']) == 4
-
-def test_load_all(fpath):
-    with dlisio.load(fpath) as (f, *_):
-        f.load()
-        assert len(f.indexedobjects) == 23
-
-def test_unknowns_loading():
-    with dlisio.load('data/206_05a-_3_DWL_DWL_WIRE_258276498.DLIS') as (f,):
-        assert len(f.indexedobjects) == 4 #FILE-HEADER, ORIGIN, FRAME, CHANNEL
-        assert len(f.unknowns)       == 5
-        assert len(f.indexedobjects) == 9
